apps/mcp-servers/debug/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finsight-mcp-debug"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "httpx>=0.27",
  "redis[hiredis]>=5.0",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "structlog>=24.1",
  "asyncpg>=0.29",
  "docker>=7.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
